update ffn.cpp and ffn.h with fp16

+++ ffn.h
@@ -1,7 +1,7 @@
 #ifndef FFN_H
 #define FFN_H
 
-void geglu(const float* x, float* out, int B);
+void geglu(const __half* x, __half* out, int B);
 void initialize_once();
 
 extern float *Wu;
